requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
pydantic
passlib
bcrypt==4.0.1
python-jose